# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CKB Agent Wallet Library API — programmable wallet for AI agents
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wallets/{id}/transfer:
    post:
      operationId: transferCkb
      tags: [transactions]
      summary: Transfer CKB to another address (safety rails enforced)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferInput"
      responses:
        "200":
          description: Transaction sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TxResult"
        "400":
          description: Bad request or safety rail violation
        "404":
          description: Wallet not found
  /wallets/{id}/sign:
    post:
      operationId: signTransaction
      tags: [transactions]
      summary: Sign a raw CKB transaction (safety rails enforced)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignTxInput"
      responses:
        "200":
          description: Signed transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TxResult"
        "400":
          description: Safety rail violation
        "404":
          description: Wallet not found
  /wallets/{id}/transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List transaction history for wallet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TxRecord"
        "404":
          description: Wallet not found

  # ── Safety Rails ──────────────────────────────────────────────────────────
  /wallets/{id}/safety:
    get:
      operationId: getSafetyRules
      tags: [safety]
      summary: Get safety rules for wallet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Safety rules
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SafetyRules"
        "404":
          description: Wallet not found
    put:
      operationId: updateSafetyRules
      tags: [safety]
      summary: Update safety rules for wallet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SafetyRulesUpdate"
      responses:
        "200":
          description: Updated rules
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SafetyRules"
        "404":
          description: Wallet not found
...
components:
  schemas:
...
    TransferInput:
      type: object
      required: [toAddress, amount]
      properties:
        toAddress:
          type: string
          description: Recipient CKB address
        amount:
          type: string
          description: Amount in shannons
        memo:
          type: string
          description: Optional memo attached as cell data
    SignTxInput:
      type: object
      required: [rawTx]
      properties:
        rawTx:
          type: object
          description: Raw CKB transaction JSON to sign
          additionalProperties: true
    TxResult:
      type: object
      required: [txHash, status]
      properties:
        txHash:
          type: string
        status:
          type: string
          enum: [pending, committed, rejected]
        message:
          type: ["string", "null"]
    TxRecord:
      type: object
      required: [id, walletId, action, status, createdAt]
      properties:
        id:
          type: integer
        walletId:
          type: integer
        action:
          type: string
        txHash:
          type: ["string", "null"]
        amountShannons:
          type: ["string", "null"]
        status:
          type: string
        details:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time

    # ── Safety Rails ────────────────────────────────────────────────────────
    SafetyRules:
      type: object
      required: [id, walletId, isActive, allowedActions, createdAt, updatedAt]
      properties:
        id:
          type: integer
        walletId:
          type: integer
        maxTransferAmountShannons:
          type: ["string", "null"]
          description: Max CKB per single transfer (shannons)
        dailySpendingLimitShannons:
          type: ["string", "null"]
          description: Max daily spending in shannons
        dailySpentShannons:
          type: string
          description: Amount spent today in shannons
        allowedActions:
          type: array
          items:
            type: string
          description: "Allowed actions: transfer, sign, fiber_open, fiber_pay, dob_mint, otx_compose"
        addressWhitelist:
          type: array
          items:
            type: string
          description: If non-empty, only these addresses can receive funds
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SafetyRulesUpdate:
      type: object
      properties:
        maxTransferAmountShannons:
          type: ["string", "null"]
        dailySpendingLimitShannons:
          type: ["string", "null"]
        allowedActions:
          type: array
          items:
            type: string
        addressWhitelist:
          type: array
          items:
            type: string
        isActive:
          type: boolean

    # ── Fiber ───────────────────────────────────────────────────────────────
...