# lib/api-zod/src/generated/types/safetyRulesUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CKB Agent Wallet Library API — programmable wallet for AI agents
 * OpenAPI spec version: 0.1.0
 */

export interface SafetyRulesUpdate {
  /** @nullable */
  maxTransferAmountShannons?: string | null;
  /** @nullable */
  dailySpendingLimitShannons?: string | null;
  allowedActions?: string[];
  addressWhitelist?: string[];
  isActive?: boolean;
}